Make view less fragile.
If you just do a get on the view it seems to fail. Also just use the
default data dict instead of moving it to some object. This makes it look
more like json than a django object.

Also make defaults for when the API fails or we run out of queries. The
default for google right now is a 100. Which if you are developing it will
go by quickly.